load_balancer/requirements.txt
fastapi
uvicorn
pika
python-dotenv
websockets
